Pipeline reorganization, e6 fixes
I reorganized the scraper pipeline so that it’s easier to reason about what happens where, and it’s possible for a validation step to replace an ArchiveTask with another task. This allows the duplicate source checker to detect the duplicate source and ask for help, rather than just failing the queue entry.

I also fixed a problem with e621 where it was including artist profile pages as source links, causing erroneous duplicates. There’s now a pattern-based deny-list for sources that come from e621, allowing me to exclude things that appear to be an artist’s profile page/user page/etc.