Make bot reply to originating message, fix queue bug
The bug was that unprocessable links in captions still got a score of 1, which meant that they would be used over an image, even though there’s no way for the unprocessable link to end up working. This adds more nuance to the logic to enable smarter decisions about what to do when there’s both a photo and a link.