Add database migration utility
Fullmoon conveniently provides an `upgrade()` function that does schema updates, so I buckled down and wrote the code necessary to use that on all of the user databases.  I also refactored some things so that it’s easier to do this kind of update.

Oh, and I tested and tweaked the database schema changes from the last commit, and wrote an additional migration tool to compute the image sizes.

I had to patch Fullmoon to make the dryrun mode actually do a dry run and fix a bug that caused my `order` column to make the migration fail. (It needed more quotes.)