Add bulk tag rule creation features
Now it’s possible to use the tag data that’s been saved invisibly from the scrapers. When editing an image, users can now select unused tags from the scrapers and make new tag rules easily! When a new tag rule is created this way, it will also look for any tag updates that should be made based on this new rule.

I also noticed that I broke thumbnails in the image group edit dialog before, so I fixed that.