Hack around iOS Safari’s broken CSS system colors
I’m supposed to be able to say that the theme color is `Canvas`, and have it automatically adjust to light/dark mode. It does not, but only on iOS. I’m making the experience worse for everyone except iOS users in order to get rid of the ugly white bar at the top of the screen in dark mode. Not sure what to do about the image borders yet.