Use new gradient hash impl. to check for duplicates
This involved restructuring part of the queue processing pipeline in order to make the data flow more sensible. Now, images and thumbnails are downloaded in a separate stage of the pipeline before all of the database operations start to happen, which makes it possible to check for duplicates based on the image contents.

Additionally:
- Process queue entries from oldest to newest, to preserve comic order when forwarding to the bot
- Isolate each user’s queue processing task behind pcall() so that if on user enqueues something that breaks the pipeline, only their account suffers.
- Update unit tests (the archive key is now fetch for that phase of the pipeline)