Add share log, avoid duplicate shares
Turns out that implementing the share log is a good way to prevent duplicate shares! I used a pattern similar to a CSRF token to ensure that subsequent requests to share an image with the same token will fail. This is a lousy user experience, but I make up for it by (visually) disabling the send button and preventing the form from submitting with JavaScript. (I can’t do anything about the problem for users without JavaScript, and the 400 error is enough, I think.)