Compute form submit path in lua, pass to template
Calling `makeUrl()` a bunch of times in every template was a poorly-considered hack. Deciding the submit path for the form once in the web response handler is much more reasonable. Also, it doesn’t end up generating HTTP URLs for HTTPS resources when the application is running behind e.g. nginx.