Add “discarded” status to queue, fix subtle bug
Queue items can now be “discarded” instead of just an error with a special message. This makes it easier to do the cleanup.

Also, the statuses and the error codes which set them were not aligned, so a retry error could accidentally cause the item to show up as archived even though it wasn’t.