Add support for scraping public Telegram channels
This wasn’t as hard as I was afraid it would be! Unfortunately, Telegram doesn’t include video size information in their embed webpages at all, which means I have to work around that in the record view template :(

Up next: save enough information when enqueuing via Telegram to be able to fall back to archiving by scraping the Telegram channel that a post came from. This way if an artist posts alternate versions of an image on Telegram, it’s possible to automatically archive those.