Don’t leave 0-byte media files on disk
When the disk fills up, the previous code left 0-byte image files on the disk, created by the scraper pipeline, but unable to be written into. This error also wasn’t reported to the rest of the pipeline. I’ve now corrected the oversight so that 0-byte files aren’t left laying around, and the queue entries are marked as failing.