Fix perpetually-expanding “added tags” list
Because I was using `INSERT OR IGNORE` instead of `INSERT OR REPLACE`, the inserts would fail and be ignored, preventing anything from ending up in the database.

Additionally, I resolved an issue that was causing previously-deleted tags to reappear based on tag rules. Based on Scale’s suggestion, I modified the tag rule system to only apply tags for the rules that were just created, not for every rule. I think this should be more in line with user expectations.