Add feature to disable pings for specific handles temporarily
This is most useful when someone is having a bad time and wouldn’t appreciate being mentioned for shared art.

It was difficult to dive into the thorniest part of the code (the SPL edit page) and add this. While I was there, I also realized the username/nickname fields were editable, but that wasn’t saved anywhere. I fixed that while adding the above feature.