Improve queue help UX
When someone clicks a button at the bottom of the queue help page with JS enabled, use htmx to reload the body (and remove the dialog box) without changing the scroll position. This prevents the user from needing to go find where they were agin, which is annoying (especially on mobile where that might involve quite a bit of scrolling).