Fix corrupt queue state problem caused by race
When the scraper pipeline was working on a queue item at the same time as the user was clicking things in the UI, it was possible for the prior code to create a situation where the three columns in the queue table which control state (state, help_ask, help_answer) were in an invalid combination of values. I now explicitly set the status on every modification of the queue row, which means that this kind of corruption is no longer possible.

I also spent a bunch of time building test infrastructure for proving that changes like this actually work. I didn’t test the race condition fix because I’m confident that the change I made will solve the problem. However, I noticed a different bug which caused the `RetryLimitReached` error state to never actually happen, so I did write a test for that! In the process, I fixed some confusingly named variables and improved the robustness of the scraper tests.