Remove TLS from nginx to Werehouse
Previously, Redbean wouldn't allow me to set __Host- or __Secure-
cookies unless it was doing the TLS.  I updated Redbean to allow those
cookie name prefixes when the client is a trusted IP and the request
includes an X-Forwarded-Proto: https header.  Now the flake doesn't need
to configure nginx to use TLS when talking to Werehouse, which improves
performance.  (The TLS wasn't necessary anyway; they're talking to each
other over a loopback interface.)